tasks-api/dbInitLambda/requirements.txt
boto3
psycopg2-binary
urllib3